requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
Pillow
requests
streamlit
python-dotenv
